+++ changelog.md
@@ -3,6 +3,8 @@
 * Portals now generate their target structure rather than searching for it, reducing initial loading time
 * Generated structures are now less likely to be buried
 * Fixed Sus sand/gravels loot tables
+* Fixed sounds not sounding
+* Increased portal cooldown to avoid unexpected back teleportation
 
 # 0.0.1